+++ src/main/java/carRacing/CarRacingGame.java
@@ -1,4 +1,4 @@
-package step3;
+package carRacing;
 
 import java.util.ArrayList;
 import java.util.List;
