+++ src/main/java/carRacing/InputView.java
@@ -1,4 +1,4 @@
-package step3;
+package carRacing;
 
 import java.util.Scanner;
