+++ src/main/java/carRacing/ResultView.java
@@ -1,4 +1,4 @@
-package step3;
+package carRacing;
 
 import java.util.List;
 import java.util.stream.IntStream;
