+++ src/main/java/carRacing/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package step2;
+package carRacing;
 
 import java.util.Arrays;
 import java.util.regex.Matcher;
