+++ src/test/java/carRacing/CarTest.java
@@ -1,5 +1,6 @@
-package step3;
+package carRacing;
 
+import carRacing.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
