+++ src/test/java/carRacing/RandomUtilTest.java
@@ -1,4 +1,4 @@
-package step3;
+package carRacing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
