+++ src/test/java/carRacing/SetTest.java
@@ -1,4 +1,4 @@
-package step1;
+package carRacing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
