+++ src/test/java/carRacing/StringTest.java
@@ -1,4 +1,4 @@
-package step1;
+package carRacing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
